[skip ci] Human vs mouse synteny
