+++ products/jbrowse-web/src/Loader.tsx
@@ -43,6 +43,7 @@ function NoConfigMessage() {
     ['test_data/breakpoint/config.json', 'Breakpoint'],
     ['test_data/config_dotplot.json', 'Grape/Peach dotplot'],
     ['test_data/config_synteny_grape_peach.json', 'Grape/Peach synteny'],
+    ['test_data/human_vs_mouse.json', 'Human vs mouse synteny'],
     ['test_data/yeast_synteny/config.json', 'Yeast synteny'],
     ['test_data/config_many_contigs.json', 'Many contigs'],
     ['test_data/config_honeybee.json', 'Honeybee'],
